refactoring flickity into its own module.
This solves the js error as well.
