+++ src/components/ArtObjectPageComponents/TabbedSubMenu/index.jsx
@@ -3,13 +3,10 @@ import { Link } from 'react-router-dom'
 import PanelVisuallyRelated from '../PanelVisuallyRelated'
 import PanelEnsemble from '../PanelEnsemble'
 import PanelDetails from '../PanelDetails'
+import FlickityMenu from './FlickityMenu'
 import { getArtObjectUrlFromId } from '../../../helpers'
 import MediaQuery from 'react-responsive'
 import { BREAKPOINTS } from '../../../constants'
-
-// note: #flickityWorkaround link directly to the js, because otherwise flickity requires you to change the webpack config,
-// and react-scripts doesn't let you do that without "ejecting".
-const Flickity = require('../../../../node_modules/flickity/dist/flickity.pkgd.js')
 
 class TabbedSubMenu extends Component {
   constructor(props) {
@@ -33,35 +30,6 @@ class TabbedSubMenu extends Component {
     this.state = { tabs: tabList }
   }
 
-  componentDidMount() {
-    // this should always find a value
-    const selectedIndex = this.state.tabs.map(function(tab) {
-      return tab.slug
-    }).indexOf(this.props.slug)
-
-    const options = {
-      // align the first one to the left, otherwise, center it.
-      cellAlign: selectedIndex === 0 ? 'left' : 'center',
-      initialIndex: selectedIndex,
-      accessibility: true,
-      pageDots: false,
-      prevNextButtons: false,
-      wrapAround: false,
-    }
-
-    // note: #flickityWorkaround - because we can't use react-flickity without changing webpack configs
-    this.flickityInstance = new Flickity( '.tabs-flickity', options)
-  }
-
-  componentWillUnMount() {
-    // note: #flickityWorkaround - because we can't use react-flickity without changing webpack configs
-    if (this.flickityInstance) {
-      this.flickityInstance.destroy()
-    }
-
-    this.flickityInstance = null
-  }
-
   getTab() {
     switch(this.props.slug) {
       case 'ensemble':
@@ -76,6 +44,11 @@ class TabbedSubMenu extends Component {
   render() {
     const ensembleIsDisabled = !this.props.object.ensembleIndex
     const props = this.props
+
+    const selectedIndex = this.state.tabs.map(function(tab) {
+      return tab.slug
+    }).indexOf(this.props.slug)
+
     const tabsList = this.state.tabs.map(tabData => {
       const isSelected = tabData.slug === props.slug
       return (
@@ -104,9 +77,10 @@ class TabbedSubMenu extends Component {
         <div className="container">
           <nav className="m-tabs m-tabs--post-cta" data-behavior="Tabs">
             <MediaQuery maxWidth={BREAKPOINTS.tablet_max}>
-              <div className="m-tabs__list tabs-list-mobile tabs-flickity">
-                {tabsList}
-              </div>
+              <FlickityMenu
+                tabsList={tabsList}
+                selectedIndex={selectedIndex}
+              />
             </MediaQuery>
             <MediaQuery minWidth={BREAKPOINTS.tablet_max + 1}>
               <div className="m-tabs__list tabs-list-desktop">
